# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Travel Assistant AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ChatMessageInput {
  message: string;
  sessionId?: string;
}

export type McpLogEntryType = typeof McpLogEntryType[keyof typeof McpLogEntryType];


export const McpLogEntryType = {
  connect: 'connect',
  init: 'init',
  tool_call: 'tool_call',
  tool_args: 'tool_args',
  tool_response: 'tool_response',
  error: 'error',
  info: 'info',
} as const;

/**
 * @nullable
 */
export type McpLogEntryData = { [key: string]: unknown } | null;

export interface McpLogEntry {
  timestamp: string;
  type: McpLogEntryType;
  service: string;
  content: string;
  /** @nullable */
  data?: McpLogEntryData;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: string;
  role: ChatMessageRole;
  content: string;
  timestamp: string;
  mcpLogs?: McpLogEntry[];
}

export interface ChatResponse {
  message: ChatMessage;
  mcpLogs: McpLogEntry[];
}

export type GetChatHistoryParams = {
sessionId?: string;
};

export type ClearChatHistoryParams = {
sessionId?: string;
};