# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Travel Assistant AI API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: chat
    description: AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /chat:
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message to the AI travel assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response with MCP logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /chat/history:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get chat history for a session
      parameters:
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

    delete:
      operationId: clearChatHistory
      tags: [chat]
      summary: Clear chat history for a session
      parameters:
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: History cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    ChatMessageInput:
      type: object
      properties:
        message:
          type: string
        sessionId:
          type: string
      required:
        - message

    McpLogEntry:
      type: object
      properties:
        timestamp:
          type: string
        type:
          type: string
          enum: [connect, init, tool_call, tool_args, tool_response, error, info]
        service:
          type: string
        content:
          type: string
        data:
          type: ["object", "null"]
      required:
        - timestamp
        - type
        - service
        - content

    ChatMessage:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        timestamp:
          type: string
        mcpLogs:
          type: array
          items:
            $ref: "#/components/schemas/McpLogEntry"
      required:
        - id
        - role
        - content
        - timestamp

    ChatResponse:
      type: object
      properties:
        message:
          $ref: "#/components/schemas/ChatMessage"
        mcpLogs:
          type: array
          items:
            $ref: "#/components/schemas/McpLogEntry"
      required:
        - message
        - mcpLogs