# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Travel Assistant AI API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';
import type { McpLogEntry } from './mcpLogEntry';

export interface ChatResponse {
  message: ChatMessage;
  mcpLogs: McpLogEntry[];
}