# lib/api-zod/src/generated/types/mcpLogEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Travel Assistant AI API
 * OpenAPI spec version: 0.1.0
 */
import type { McpLogEntryData } from './mcpLogEntryData';
import type { McpLogEntryType } from './mcpLogEntryType';

export interface McpLogEntry {
  timestamp: string;
  type: McpLogEntryType;
  service: string;
  content: string;
  /** @nullable */
  data?: McpLogEntryData;
}